refactor(chat-thread): split the windowed body out and fix the handoff
The virtualization PR grew ChatThread.tsx to ~1900 lines and react-doctor
flagged the windowed body as a giant component pushing live state to its
parent through a ref-filled effect.

- Move the windowed body to VirtualThreadScrollBody.tsx, decomposed into
  useReservedFooterHeight / useSettleControls / useStickyAnchor /
  useFollowBottom. Row rendering comes in as a `renderRow` prop, so the
  row markup stays in ChatThread and there is no import cycle.
- Move the sticky headers to ThreadStickyHeader.tsx and make the
  non-virtualized overlay clear its dismissal in render, like the
  windowed variant already did, instead of chaining through an effect.
- Replace the jump-API ref + effect with a `renderNav` render prop: the
  nav layer is now composed by the parent but rendered beside the
  scroller, so the body hands it the jump implementation directly.
- Fix the mid-session handoff. The two bodies do not share a scroll
  coordinate space (virtual offsets are 80px estimates until rows
  measure), so resuming from a recorded `scrollTop` landed anywhere,
  usually clamped to the bottom. Record the engine's anchored user
  message instead and resume with scrollToIndex. That also drops the
  recorder's scroll listener and DOM probe.
- Add a dev-only Advanced setting overriding the virtualization
  threshold, so the windowed path and the mid-session flip can be tested
  without a 150-row conversation, plus a dev-only amber ring and
  "windowed mounted/total" badge on the windowed body — the two
  renderers are otherwise pixel-identical.
