+++ packages/ui/src/features/sessions/components/chat-thread/ChatThread.tsx
@@ -1,6 +1,5 @@
 import {
   CaretDown,
-  ChatCircle,
   Check,
   Copy,
   FileText,
@@ -9,7 +8,6 @@ import {
 import { WorkerPoolContextProvider } from "@pierre/diffs/react";
 import { useService } from "@posthog/di/react";
 import {
-  Button,
   ChatBubble,
   ChatBubbleContent,
   ChatMarker,
@@ -48,22 +46,23 @@ import {
   type PromptRecallHandler,
 } from "@posthog/ui/features/sessions/components/chat-thread/composerPromptRecall";
 import { MessageJumpPicker } from "@posthog/ui/features/sessions/components/chat-thread/MessageJumpPicker";
+import { StickyHeaderOverlay } from "@posthog/ui/features/sessions/components/chat-thread/ThreadStickyHeader";
 import { ToolGroup } from "@posthog/ui/features/sessions/components/chat-thread/ToolGroup";
 import { THREAD_HOTKEY_OPTIONS } from "@posthog/ui/features/sessions/components/chat-thread/threadHotkeys";
 import {
   type AgentTurn,
   CHAT_THREAD_VIRTUALIZATION_THRESHOLD,
   completedTurnTimestamp,
-  computeStickyAnchor,
   countFlatRows,
   type FlatThreadRow,
   flattenTurnRows,
-  type StickyAnchorEntry,
-  type StickyAnchorState,
+  SCROLL_PREVIOUS_ITEM_PEEK,
   type ThreadItem,
+  type ThreadScrollResume,
   type TurnRow,
 } from "@posthog/ui/features/sessions/components/chat-thread/threadVirtualization";
 import { usePromptRecallSource } from "@posthog/ui/features/sessions/components/chat-thread/usePromptRecallSource";
+import { VirtualThreadScrollBody } from "@posthog/ui/features/sessions/components/chat-thread/VirtualThreadScrollBody";
 import { GitActionMessage } from "@posthog/ui/features/sessions/components/GitActionMessage";
 import { GitActionResult } from "@posthog/ui/features/sessions/components/GitActionResult";
 import { mergeConversationItems } from "@posthog/ui/features/sessions/components/mergeConversationItems";
@@ -101,11 +100,8 @@ import {
   type DiffWorkerFactory,
 } from "@posthog/ui/shell/diffWorkerHost";
 import { IconButton, Tooltip } from "@radix-ui/themes";
-import { useVirtualizer } from "@tanstack/react-virtual";
-import { AnimatePresence, motion, useReducedMotion } from "framer-motion";
 import {
   memo,
-  type MouseEvent as ReactMouseEvent,
   type ReactNode,
   type RefObject,
   useCallback,
@@ -118,12 +114,6 @@ import {
 import { useHotkeys } from "react-hotkeys-hook";
 
 type SessionUpdateItem = Extract<ConversationItem, { type: "session_update" }>;
-
-/**
- * How far below the viewport top a user message may sit while still counting as the current
- * anchor — shared by the engine (`scrollPreviousItemPeek`) and the virtualized sticky header.
- */
-const SCROLL_PREVIOUS_ITEM_PEEK = 64;
 
 function isToolCallItem(item: ConversationItem): item is SessionUpdateItem {
   return (
@@ -480,175 +470,6 @@ function MessageCopyButton({
 }
 
 /**
- * "Fake sticky" header. A real `position: sticky` row can't hand off in this flat list (every row
- * shares one containing block, so they'd pile at the top) and sticking causes reflow. Instead we
- * overlay a single header, out of flow, pinned over the viewport top — showing the current turn's
- * user message (the engine's anchor) once the real one has scrolled off. Click to scroll back to it.
- *
- * Only this small component subscribes to the engine's per-scroll visibility state, so the rows
- * themselves never re-render on scroll.
- */
-function StickyHeaderOverlay({ items }: { items: ConversationItem[] }) {
-  const { currentAnchorId } = useChatMessageScrollerVisibility();
-  const { scrollToMessage } = useChatMessageScroller();
-  const [dismissedId, setDismissedId] = useState<string | null>(null);
-  const [offscreen, setOffscreen] = useState(false);
-  // Anchor element used only to locate the enclosing scroller/viewport in the DOM.
-  const probeRef = useRef<HTMLSpanElement>(null);
-
-  const active = items.find(
-    (i): i is Extract<ConversationItem, { type: "user_message" }> =>
-      i.id === currentAnchorId && i.type === "user_message",
-  );
-  const activeId = active?.id ?? null;
-
-  // The engine's `visibleMessageIds` can't be used here: its IntersectionObserver excludes a band of
-  // `scrollPreviousItemPeek` px at the viewport top, which is exactly where a freshly-anchored turn
-  // message lands — so it reads as "not visible" while plainly on screen. Measure real geometry
-  // instead: the message is off-screen only once its bottom scrolls above the viewport top.
-  useEffect(() => {
-    // No reset when there's no anchor: the overlay render already guards on `active != null`, so a
-    // stale `offscreen` is never shown, and a fresh anchor re-measures synchronously below. (Avoids
-    // the prop-sync-in-effect pattern react-doctor flags.)
-    if (activeId == null) return;
-    const viewport = probeRef.current
-      ?.closest('[data-slot="chat-message-scroller"]')
-      ?.querySelector('[data-slot="chat-message-scroller-viewport"]');
-    if (!viewport) return;
-
-    const measure = () => {
-      const el = viewport.querySelector(
-        `[data-message-id="${CSS.escape(activeId)}"]`,
-      );
-      if (!el) {
-        setOffscreen(false);
-        return;
-      }
-      const messageBottom = el.getBoundingClientRect().bottom;
-      const viewportTop = viewport.getBoundingClientRect().top;
-      setOffscreen(messageBottom <= viewportTop + 4);
-    };
-
-    measure();
-    viewport.addEventListener("scroll", measure, { passive: true });
-    return () => viewport.removeEventListener("scroll", measure);
-  }, [activeId]);
-
-  // Once the real message is back on screen, clear the dismissal so the header can return later.
-  useEffect(() => {
-    if (!offscreen) setDismissedId(null);
-  }, [offscreen]);
-
-  const dismiss = (id: string) => {
-    // Hide immediately on click (don't wait for the scroll to bring the message into view), then
-    // jump to it.
-    setDismissedId(id);
-    scrollToMessage(id);
-  };
-
-  return (
-    <>
-      <span ref={probeRef} className="hidden" aria-hidden="true" />
-      <AnimatePresence>
-        {active != null && offscreen && active.id !== dismissedId && (
-          <StickyHeaderJumpButton onClick={() => dismiss(active.id)} />
-        )}
-      </AnimatePresence>
-    </>
-  );
-}
-
-/**
- * The floating "jump to your message" pill both sticky-header variants render. Must be a direct
- * child of an `AnimatePresence` so its exit animation plays.
- */
-function StickyHeaderJumpButton({ onClick }: { onClick: () => void }) {
-  const shouldReduceMotion = useReducedMotion();
-  return (
-    <motion.div
-      key="chat-sticky-header"
-      // Slide in slightly from the top + fade (ease-out-cubic). Exit a touch faster.
-      initial={shouldReduceMotion ? false : { opacity: 0, y: -8 }}
-      animate={{ opacity: 1, y: 0 }}
-      exit={
-        shouldReduceMotion
-          ? { opacity: 0 }
-          : { opacity: 0, y: -8, transition: { duration: 0.15 } }
-      }
-      transition={{ duration: 0.2, ease: [0.215, 0.61, 0.355, 1] }}
-      // pointer-events-none on the strip so only the button catches clicks — the rest stays
-      // transparent to the content scrolling underneath.
-      className="pointer-events-none absolute inset-x-0 top-2 z-10"
-    >
-      {/* Align to the content column's right edge (matches the message rows) rather than the
-          viewport edge, so the button reads in-context with the conversation. */}
-      <div
-        className="mx-auto flex w-full justify-end px-2"
-        style={{ maxWidth: CHAT_CONTENT_MAX_WIDTH }}
-      >
-        <Button
-          type="button"
-          variant="outline"
-          size="icon"
-          title="Jump to your message"
-          aria-label="Jump to your message"
-          onClick={onClick}
-          className="pointer-events-auto rounded-full bg-background shadow-md"
-        >
-          <ChatCircle />
-        </Button>
-      </div>
-    </motion.div>
-  );
-}
-
-/**
- * Virtualized-mode sticky header. The engine's visibility state can't power it here — unmounted
- * rows aren't observed — so the parent derives the anchor from the virtualizer's measurements
- * (see {@link computeStickyAnchor}) and passes the result down. Dismissal semantics match
- * {@link StickyHeaderOverlay}: hide immediately on click, return once the message has been back
- * on screen.
- */
-function VirtualStickyHeader({
-  items,
-  anchorId,
-  offscreen,
-  onJump,
-}: {
-  items: ConversationItem[];
-  anchorId: string | null;
-  offscreen: boolean;
-  onJump: (id: string) => void;
-}) {
-  const [dismissedId, setDismissedId] = useState<string | null>(null);
-
-  // Clear a dismissal the moment the anchor is back on screen, so the header can return for the
-  // next offscreen episode. Render-phase adjustment, not an effect — the cleared state must not
-  // flash through a committed frame.
-  if (!offscreen && dismissedId !== null) {
-    setDismissedId(null);
-  }
-
-  const active = items.find(
-    (i): i is Extract<ConversationItem, { type: "user_message" }> =>
-      i.id === anchorId && i.type === "user_message",
-  );
-
-  return (
-    <AnimatePresence>
-      {active != null && offscreen && active.id !== dismissedId && (
-        <StickyHeaderJumpButton
-          onClick={() => {
-            setDismissedId(active.id);
-            onJump(active.id);
-          }}
-        />
-      )}
-    </AnimatePresence>
-  );
-}
-
-/**
  * Start-aligned assistant prose bubble. Streamed tokens arrive in bursts; `useSmoothedText` reveals
  * them at a steady character rate so the text reads as even typing (text present on mount shows
  * immediately, so completed messages render in full with no replay).
@@ -973,17 +794,12 @@ function ThreadKeyboardNav({
 }
 
 /**
- * Scroll state the non-virtualized body continuously records so the virtualized body can resume
- * from roughly the same place when the thread crosses the virtualization threshold mid-session.
- */
-type ThreadScrollResume = { atBottom: boolean; scrollTop: number };
-
-/**
- * Keeps {@link ThreadScrollResume} current while the non-virtualized body is mounted. At-bottom
- * comes from the engine's scrollable state (`end` is true while content extends below the fold);
- * scrollTop comes from a passive listener on the engine viewport, located via the same hidden
- * probe pattern the other engine-adjacent helpers use. Writes go to a ref — this must not
- * re-render on scroll.
+ * Keeps {@link ThreadScrollResume} current while the non-virtualized body is mounted, so the
+ * windowed body can pick up where this one left off if the thread crosses the threshold
+ * mid-session. Both values come from engine state the scroller already tracks — at-bottom from
+ * `scrollable.end` (true while content extends below the fold), and the anchored user message from
+ * the same visibility state the sticky header reads — so nothing here listens to scroll. Writes go
+ * to a ref: the recorder must never make the thread re-render.
  */
 function ThreadScrollStateRecorder({
   stateRef,
@@ -991,26 +807,13 @@ function ThreadScrollStateRecorder({
   stateRef: RefObject<ThreadScrollResume>;
 }) {
   const { end } = useChatMessageScrollerScrollable();
-  const probeRef = useRef<HTMLSpanElement>(null);
+  const { currentAnchorId } = useChatMessageScrollerVisibility();
 
   useEffect(() => {
-    stateRef.current = { ...stateRef.current, atBottom: !end };
-  }, [end, stateRef]);
+    stateRef.current = { atBottom: !end, anchorId: currentAnchorId ?? null };
+  }, [end, currentAnchorId, stateRef]);
 
-  useEffect(() => {
-    const viewport = probeRef.current
-      ?.closest('[data-slot="chat-message-scroller"]')
-      ?.querySelector('[data-slot="chat-message-scroller-viewport"]');
-    if (!viewport) return;
-    const record = () => {
-      stateRef.current = { ...stateRef.current, scrollTop: viewport.scrollTop };
-    };
-    record();
-    viewport.addEventListener("scroll", record, { passive: true });
-    return () => viewport.removeEventListener("scroll", record);
-  }, [stateRef]);
-
-  return <span ref={probeRef} className="hidden" aria-hidden="true" />;
+  return null;
 }
 
 /** The scroll body, under the Provider so the overlay + scroll-button hooks can read engine state. */
@@ -1181,24 +984,7 @@ function ThreadScrollbarRail({
   );
 }
 
-// Windowing geometry for the virtualized body. Estimate/overscan/drift values match the legacy
-// VirtualizedList, whose tuning these rows share (same item mix, same measure-then-settle churn).
-const VIRTUAL_ESTIMATED_ROW_SIZE = 80;
-const VIRTUAL_OVERSCAN = 12;
-/** Matches the Provider's `scrollEdgeThreshold` so "at bottom" agrees between engine and windowing. */
-const VIRTUAL_AT_BOTTOM_THRESHOLD = 100;
-// A real upward drift, not a 1-frame measure transient: the DOM bottom sits
-// this far below the viewport. Well above any single append's measure gap.
-const VIRTUAL_FAR_DRIFT_THRESHOLD = 400;
-/** Top of the virtual coordinate space — stands in for the non-virtualized content's `py-4`. */
-const VIRTUAL_PADDING_START = 16;
-
 const EMPTY_FLAT_ROWS: FlatThreadRow[] = [];
-
-/** Imperative surface the virtualized body hands to the nav layer rendered outside it. */
-interface VirtualJumpApi {
-  jumpToMessage: (id: string) => void;
-}
 
 /**
  * One windowed row. Memoized against the row's *contents* rather than the row wrapper object —
@@ -1255,395 +1041,6 @@ const FlatRowView = memo(
     prev.renderItem === next.renderItem &&
     prev.keyboardFocused === next.keyboardFocused,
 );
-
-/**
- * Windowed scroll body for long threads, following the upstream MessageScroller guidance:
- * virtualization lives outside the primitive — the quill viewport stays the scroll element and a
- * `@tanstack/react-virtual` virtualizer owns the rows inside `ChatMessageScrollerContent`.
- *
- * The engine still provides the chrome that only needs scroll geometry (the scroll-to-bottom
- * button and edge state read real element measurements), while everything item-based gets a
- * windowed implementation here: follow-bottom via `anchorTo: "end"` + `followOnAppend` (the
- * legacy `VirtualizedList` recipe), message jumps via `scrollToIndex`, the sticky header via
- * {@link computeStickyAnchor} over the virtualizer's measurements, and the scrollbar rail via its
- * existing interpolate-unmounted-rows path.
- */
-function VirtualThreadScrollBody({
-  items,
-  flatRows,
-  renderItem,
-  footer,
-  keyboardFocusedMessageId,
-  onUserInteract,
-  jumpApiRef,
-  resumeStateRef,
-}: {
-  items: ConversationItem[];
-  flatRows: FlatThreadRow[];
-  renderItem: (item: ConversationItem) => ReactNode;
-  /** Status row (duration / context usage) pinned as the last item in the thread. */
-  footer?: ReactNode;
-  keyboardFocusedMessageId?: string | null;
-  /** Clears keyboard-focused message state on any pointer interaction with the thread. */
-  onUserInteract?: () => void;
-  /** Filled with this body's jump implementation for the nav layer rendered outside it. */
-  jumpApiRef: RefObject<VirtualJumpApi | null>;
-  /** Scroll state recorded by the non-virtualized body, consumed once when this body takes over. */
-  resumeStateRef: RefObject<ThreadScrollResume>;
-}) {
-  const viewportRef = useRef<HTMLDivElement | null>(null);
-  const [viewportEl, setViewportEl] = useState<HTMLDivElement | null>(null);
-  const [contentEl, setContentEl] = useState<HTMLDivElement | null>(null);
-  const footerRef = useRef<HTMLDivElement | null>(null);
-  const initializedRef = useRef(false);
-  // Seeded from the handoff state so a mid-session flip while reading above the fold doesn't
-  // immediately re-pin to the bottom.
-  const isAtBottomRef = useRef(resumeStateRef.current?.atBottom ?? true);
-  const lastScrollTopRef = useRef(0);
-  const settleRafRef = useRef<number | null>(null);
-
-  // The footer is real trailing content, NOT a fake virtual row — as a virtual row its constant
-  // key would always be last and permanently kill tanstack's followOnAppend. Its height is
-  // reserved as `paddingEnd` instead so the virtual coordinate space includes it (the
-  // VirtualizedList recipe).
-  const hasFooter = footer != null;
-  const [footerHeight, setFooterHeight] = useState(0);
-  useLayoutEffect(() => {
-    const el = footerRef.current;
-    const measure = () => {
-      const height = hasFooter && el ? el.offsetHeight : 0;
-      setFooterHeight((prev) => (prev === height ? prev : height));
-    };
-    measure();
-    if (!hasFooter || !el) return;
-    const ro = new ResizeObserver(measure);
-    ro.observe(el);
-    return () => ro.disconnect();
-  }, [hasFooter]);
-
-  const virtualizer = useVirtualizer({
-    count: flatRows.length,
-    getScrollElement: () => viewportRef.current,
-    estimateSize: () => VIRTUAL_ESTIMATED_ROW_SIZE,
-    overscan: VIRTUAL_OVERSCAN,
-    anchorTo: "end",
-    followOnAppend: true,
-    scrollEndThreshold: VIRTUAL_AT_BOTTOM_THRESHOLD,
-    paddingStart: VIRTUAL_PADDING_START,
-    paddingEnd: footerHeight,
-    getItemKey: (index) => flatRows[index]?.key ?? index,
-  });
-
-  const cancelSettle = useCallback(() => {
-    if (settleRafRef.current !== null) {
-      cancelAnimationFrame(settleRafRef.current);
-      settleRafRef.current = null;
-    }
-  }, []);
-
-  // Pin to the true bottom, retrying across frames while rows measure taller than the estimate.
-  const settleAtEnd = useCallback(() => {
-    cancelSettle();
-    isAtBottomRef.current = true;
-    let attempts = 0;
-    const step = () => {
-      virtualizer.scrollToEnd();
-      if (virtualizer.isAtEnd(VIRTUAL_AT_BOTTOM_THRESHOLD) || ++attempts > 12) {
-        settleRafRef.current = null;
-        return;
-      }
-      settleRafRef.current = requestAnimationFrame(step);
-    };
-    step();
-  }, [virtualizer, cancelSettle]);
-
-  // Jump to a row, re-issuing across a few frames until the target offset stops drifting as the
-  // rows around it measure.
-  const settleToIndex = useCallback(
-    (index: number) => {
-      cancelSettle();
-      isAtBottomRef.current = false;
-      virtualizer.scrollToIndex(index, { align: "start" });
-      let attempts = 0;
-      const step = () => {
-        settleRafRef.current = null;
-        const viewport = viewportRef.current;
-        const target = virtualizer.getOffsetForIndex(index, "start")?.[0];
-        if (!viewport || target == null) return;
-        const maxScroll = Math.max(
-          0,
-          viewport.scrollHeight - viewport.clientHeight,
-        );
-        if (
-          Math.abs(viewport.scrollTop - Math.min(target, maxScroll)) <= 1 ||
-          ++attempts > 8
-        ) {
-          return;
-        }
-        virtualizer.scrollToIndex(index, { align: "start" });
-        settleRafRef.current = requestAnimationFrame(step);
-      };
-      settleRafRef.current = requestAnimationFrame(step);
-    },
-    [virtualizer, cancelSettle],
-  );
-
-  useEffect(() => cancelSettle, [cancelSettle]);
-
-  const userRows = useMemo(() => {
-    const result: Array<{ id: string; rowIndex: number; content: string }> = [];
-    flatRows.forEach((row, index) => {
-      if (row.item.type === "user_message") {
-        result.push({
-          id: row.item.id,
-          rowIndex: index,
-          content: row.item.content,
-        });
-      }
-    });
-    return result;
-  }, [flatRows]);
-
-  const rowIndexByMessageId = useMemo(() => {
-    const map = new Map<string, number>();
-    flatRows.forEach((row, index) => {
-      map.set(row.item.id, index);
-    });
-    return map;
-  }, [flatRows]);
-
-  const jumpToMessage = useCallback(
-    (id: string) => {
-      const index = rowIndexByMessageId.get(id);
-      if (index != null) settleToIndex(index);
-    },
-    [rowIndexByMessageId, settleToIndex],
-  );
-
-  useEffect(() => {
-    jumpApiRef.current = { jumpToMessage };
-    return () => {
-      jumpApiRef.current = null;
-    };
-  }, [jumpApiRef, jumpToMessage]);
-
-  // Initial position: resume near where the non-virtualized body left off (offsets are estimates
-  // until rows measure, so the landing is approximate and self-corrects), or settle at the bottom.
-  useLayoutEffect(() => {
-    if (initializedRef.current || flatRows.length === 0) return;
-    initializedRef.current = true;
-    const resume = resumeStateRef.current;
-    if (resume && !resume.atBottom) {
-      virtualizer.scrollToOffset(resume.scrollTop);
-      return;
-    }
-    settleAtEnd();
-  }, [flatRows.length, settleAtEnd, virtualizer, resumeStateRef]);
-
-  // Sticky-header anchor, derived from the virtualizer's measurements (estimated for rows that
-  // have never mounted; exact once measured). Recomputed at most once per frame on scroll and on
-  // content growth.
-  const [stickyState, setStickyState] = useState<StickyAnchorState>({
-    anchorId: null,
-    offscreen: false,
-  });
-  const recomputeSticky = useCallback(() => {
-    const viewport = viewportRef.current;
-    if (!viewport) return;
-    const cache = virtualizer.measurementsCache;
-    const entries: StickyAnchorEntry[] = userRows.map((user) => {
-      const measured = cache[user.rowIndex];
-      const start =
-        measured?.start ?? user.rowIndex * VIRTUAL_ESTIMATED_ROW_SIZE;
-      const end = measured?.end ?? start + VIRTUAL_ESTIMATED_ROW_SIZE;
-      return { id: user.id, start, end };
-    });
-    const next = computeStickyAnchor(
-      entries,
-      viewport.scrollTop,
-      SCROLL_PREVIOUS_ITEM_PEEK,
-    );
-    setStickyState((prev) =>
-      prev.anchorId === next.anchorId && prev.offscreen === next.offscreen
-        ? prev
-        : next,
-    );
-  }, [userRows, virtualizer]);
-
-  const stickyFrameRef = useRef<number | null>(null);
-  const scheduleStickyRecompute = useCallback(() => {
-    if (stickyFrameRef.current != null) return;
-    stickyFrameRef.current = requestAnimationFrame(() => {
-      stickyFrameRef.current = null;
-      recomputeSticky();
-    });
-  }, [recomputeSticky]);
-  useEffect(() => {
-    scheduleStickyRecompute();
-    return () => {
-      if (stickyFrameRef.current != null) {
-        cancelAnimationFrame(stickyFrameRef.current);
-        stickyFrameRef.current = null;
-      }
-    };
-  }, [scheduleStickyRecompute]);
-
-  const handleScroll = useCallback(() => {
-    const el = viewportRef.current;
-    const scrollTop = el?.scrollTop ?? 0;
-    // Tolerate sub-pixel jitter; only a real upward move counts as leaving end.
-    const scrolledUp = scrollTop < lastScrollTopRef.current - 1;
-    lastScrollTopRef.current = scrollTop;
-
-    const atEnd = virtualizer.isAtEnd(VIRTUAL_AT_BOTTOM_THRESHOLD);
-    // Genuine far drift (not a 1-frame measure transient): the DOM bottom sits well below the
-    // viewport, so follow can't get silently stuck mid-thread.
-    const farFromEnd = el
-      ? el.scrollHeight - el.clientHeight - scrollTop >
-        VIRTUAL_FAR_DRIFT_THRESHOLD
-      : false;
-    // Hysteresis: each append measures taller than the estimate, so for one frame isAtEnd reads
-    // false before followOnAppend/anchorTo re-pin. Re-arm at the end; only clear on a real
-    // upward scroll or a genuine drift.
-    if (atEnd) {
-      isAtBottomRef.current = true;
-    } else if (scrolledUp || farFromEnd) {
-      isAtBottomRef.current = false;
-    }
-    scheduleStickyRecompute();
-  }, [virtualizer, scheduleStickyRecompute]);
-
-  const totalSize = virtualizer.getTotalSize();
-
-  // Anything that changes the virtual height while following has to re-pin to the new bottom:
-  // rows remeasuring past the estimate, late async content (highlighting, diffs) growing rows,
-  // and the footer resize feeding paddingEnd. totalSize is the one value that moves for all of
-  // them. Layout effect so the re-pin lands before paint.
-  // biome-ignore lint/correctness/useExhaustiveDependencies: totalSize is the trigger, not a body dependency
-  useLayoutEffect(() => {
-    if (!isAtBottomRef.current) return;
-    virtualizer.scrollToEnd();
-  }, [totalSize, virtualizer]);
-
-  // A prompt submitted from anywhere re-engages follow (same trigger ThreadAutoFollow uses in the
-  // non-virtualized body): settleAtEnd flips isAtBottomRef, and the totalSize re-pin plus
-  // followOnAppend keep the reply pinned from there.
-  const lastItem = items.at(-1);
-  const userMessageCount = useMemo(
-    () =>
-      items.reduce((n, item) => (item.type === "user_message" ? n + 1 : n), 0),
-    [items],
-  );
-  const prevUserCountRef = useRef(userMessageCount);
-  useLayoutEffect(() => {
-    const previous = prevUserCountRef.current;
-    prevUserCountRef.current = userMessageCount;
-    if (previous === 0 || userMessageCount <= previous) return;
-    if (lastItem?.type !== "user_message") return;
-    settleAtEnd();
-  }, [userMessageCount, lastItem, settleAtEnd]);
-
-  // Coming back to a backgrounded tab: ResizeObserver callbacks were throttled, so re-settle if
-  // we were following (the legacy view does the same).
-  useEffect(() => {
-    const handleVisibilityChange = () => {
-      if (!document.hidden && isAtBottomRef.current) settleAtEnd();
-    };
-    document.addEventListener("visibilitychange", handleVisibilityChange);
-    return () =>
-      document.removeEventListener("visibilitychange", handleVisibilityChange);
-  }, [settleAtEnd]);
-
-  const railUserMessages = useMemo(
-    () =>
-      userRows.map((user) => ({
-        id: user.id,
-        content: user.content,
-        index: user.rowIndex,
-      })),
-    [userRows],
-  );
-  const railMarkers = useMessageRailMarkers({
-    contentEl,
-    scrollEl: viewportEl,
-    userMessages: railUserMessages,
-    onJump: jumpToMessage,
-    activeId: keyboardFocusedMessageId,
-    rowAttribute: "data-message-id",
-  });
-
-  return (
-    <ChatMessageScroller
-      className="group/thread"
-      onPointerDownCapture={onUserInteract}
-    >
-      <VirtualStickyHeader
-        items={items}
-        anchorId={stickyState.anchorId}
-        offscreen={stickyState.offscreen}
-        onJump={jumpToMessage}
-      />
-      <MessageScrollbarRail markers={railMarkers} />
-      <ChatMessageScrollerViewport
-        ref={(el: HTMLDivElement | null) => {
-          viewportRef.current = el;
-          setViewportEl(el);
-        }}
-        onScroll={handleScroll}
-      >
-        {/* `block` overrides the content's flex+gap layout — spacing moves into the rows
-            (pb-4) and the virtual paddings, so translateY offsets are the whole layout. */}
-        <ChatMessageScrollerContent className="block" density="default">
-          <div
-            ref={setContentEl}
-            className="relative w-full"
-            style={{ height: totalSize }}
-          >
-            {virtualizer.getVirtualItems().map((virtualItem) => {
-              const row = flatRows[virtualItem.index];
-              if (!row) return null;
-              return (
-                <div
-                  key={virtualItem.key}
-                  ref={virtualizer.measureElement}
-                  data-index={virtualItem.index}
-                  className="absolute top-0 left-0 w-full"
-                  style={{ transform: `translateY(${virtualItem.start}px)` }}
-                >
-                  <FlatRowView
-                    row={row}
-                    renderItem={renderItem}
-                    keyboardFocused={row.item.id === keyboardFocusedMessageId}
-                  />
-                </div>
-              );
-            })}
-            {/* Footer occupies the reserved paddingEnd region at the very bottom of the virtual
-                space, so the DOM bottom == the virtual end. */}
-            {hasFooter && (
-              <div ref={footerRef} className="absolute bottom-0 left-0 w-full">
-                <div
-                  className="mx-auto w-full px-2.5 pb-8"
-                  style={{ maxWidth: CHAT_CONTENT_MAX_WIDTH }}
-                >
-                  {footer}
-                </div>
-              </div>
-            )}
-          </div>
-        </ChatMessageScrollerContent>
-      </ChatMessageScrollerViewport>
-      <ChatMessageScrollerButton
-        onClick={(event: ReactMouseEvent) => {
-          // The engine's own scroll-to-end targets the current scrollHeight, which is an
-          // estimate until every row between here and the bottom has measured — settle instead.
-          event.preventDefault();
-          settleAtEnd();
-        }}
-      />
-    </ChatMessageScroller>
-  );
-}
 
 /**
  * Thread renderer built on the ChatX (quill) primitives.
@@ -1742,29 +1139,27 @@ function ChatThreadRenderer({
   // stays there for the life of this mount (see CHAT_THREAD_VIRTUALIZATION_THRESHOLD). Long
   // sessions start virtualized from the first render; a live session flips once mid-stream,
   // resuming from the scroll state the non-virtualized body recorded.
-  const flatCount = useMemo(() => countFlatRows(rows), [rows]);
-  const [virtualized, setVirtualized] = useState(
-    () => flatCount > CHAT_THREAD_VIRTUALIZATION_THRESHOLD,
+  //
+  // The threshold is overridable from Advanced settings in dev builds: lowering it below an open
+  // thread's row count flips that thread on the spot, which is how the handoff gets exercised
+  // without a 150-row conversation. Raising it back doesn't un-flip — reopen the session.
+  const thresholdOverride = useSettingsStore(
+    (s) => s.chatThreadVirtualizationThreshold,
   );
-  if (!virtualized && flatCount > CHAT_THREAD_VIRTUALIZATION_THRESHOLD) {
+  const threshold = thresholdOverride ?? CHAT_THREAD_VIRTUALIZATION_THRESHOLD;
+  const flatCount = useMemo(() => countFlatRows(rows), [rows]);
+  const [virtualized, setVirtualized] = useState(() => flatCount > threshold);
+  if (!virtualized && flatCount > threshold) {
     setVirtualized(true);
   }
   const flatRows = useMemo(
     () => (virtualized ? flattenTurnRows(rows) : EMPTY_FLAT_ROWS),
     [virtualized, rows],
   );
-  const virtualJumpApiRef = useRef<VirtualJumpApi | null>(null);
   const threadResumeRef = useRef<ThreadScrollResume>({
     atBottom: true,
-    scrollTop: 0,
+    anchorId: null,
   });
-  const jumpToMessage = useMemo(
-    () =>
-      virtualized
-        ? (id: string) => virtualJumpApiRef.current?.jumpToMessage(id)
-        : undefined,
-    [virtualized],
-  );
 
   const [jumpPickerOpen, setJumpPickerOpen] = useState(false);
   const [keyboardFocusedMessageId, setKeyboardFocusedMessageId] = useState<
@@ -1847,6 +1242,31 @@ function ChatThreadRenderer({
     />
   );
 
+  const renderWindowedRow = useCallback(
+    (row: FlatThreadRow) => (
+      <FlatRowView
+        row={row}
+        renderItem={renderItem}
+        keyboardFocused={row.item.id === keyboardFocusedMessageId}
+      />
+    ),
+    [renderItem, keyboardFocusedMessageId],
+  );
+
+  // The nav layer sits beside the scroll body so it can be handed the windowed body's jump
+  // implementation — the engine's `scrollToMessage` only reaches mounted rows.
+  const renderNav = (jumpToMessage?: (id: string) => void) => (
+    <ThreadKeyboardNav
+      items={items}
+      jumpPickerOpen={jumpPickerOpen}
+      setJumpPickerOpen={setJumpPickerOpen}
+      keyboardFocusedMessageId={keyboardFocusedMessageId}
+      setKeyboardFocusedMessageId={setKeyboardFocusedMessageId}
+      promptRecallRef={promptRecallRef}
+      jumpToMessage={jumpToMessage}
+    />
+  );
+
   return (
     <WorkerPoolContextProvider
       poolOptions={diffsPoolOptions}
@@ -1870,33 +1290,27 @@ function ChatThreadRenderer({
               <VirtualThreadScrollBody
                 items={items}
                 flatRows={flatRows}
-                renderItem={renderItem}
+                renderRow={renderWindowedRow}
                 keyboardFocusedMessageId={keyboardFocusedMessageId}
                 onUserInteract={clearKeyboardFocus}
                 footer={footer}
-                jumpApiRef={virtualJumpApiRef}
-                resumeStateRef={threadResumeRef}
+                renderNav={renderNav}
+                resumeRef={threadResumeRef}
               />
             ) : (
-              <ThreadScrollBody
-                items={items}
-                rows={rows}
-                renderItem={renderItem}
-                keyboardFocusedMessageId={keyboardFocusedMessageId}
-                onUserInteract={clearKeyboardFocus}
-                footer={footer}
-                resumeStateRef={threadResumeRef}
-              />
+              <>
+                <ThreadScrollBody
+                  items={items}
+                  rows={rows}
+                  renderItem={renderItem}
+                  keyboardFocusedMessageId={keyboardFocusedMessageId}
+                  onUserInteract={clearKeyboardFocus}
+                  footer={footer}
+                  resumeStateRef={threadResumeRef}
+                />
+                {renderNav()}
+              </>
             )}
-            <ThreadKeyboardNav
-              items={items}
-              jumpPickerOpen={jumpPickerOpen}
-              setJumpPickerOpen={setJumpPickerOpen}
-              keyboardFocusedMessageId={keyboardFocusedMessageId}
-              setKeyboardFocusedMessageId={setKeyboardFocusedMessageId}
-              promptRecallRef={promptRecallRef}
-              jumpToMessage={jumpToMessage}
-            />
           </ChatMessageScrollerProvider>
         </ChatThreadChromeProvider>
       </SessionTaskIdProvider>
